requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0
google-generativeai>=0.8.0
websockets>=12.0
